pyproject.toml
[tool.poetry]
name = "weather-bot"
version = "0.1.0"
description = ""
authors = ["Dontgoingforyou <mishael000@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
python-telegram-bot = "^21.6"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
aiocache = "^0.12.3"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
